Drop DMX DDX
It turns out xdmx currently crashes when any client attempts to use GL
and it has been in such state for about 14 years. There was a patch to
fix the problem [1] 4 years ago, but it never got merged. The last
activity on any bugs referring to xdmx has been more than 4 years ago.

Given such situation, I find it unlikely that anyone is still using xdmx
and just having the code is a drain of resources.

[1]: https://lists.x.org/archives/xorg-devel/2017-June/053919.html

Signed-off-by: Povilas Kanapickas <povilas@radix.lt>