xwayland: Fix leak of xwl_screen on init
On screen init, if any of the private type registration fails we would
return FALSE without actually freeing the xwl_screen we just allocated.

This is not a serious leak as failure at that point would lead to the
premature termination of Xwayland at startup, but covscan complains and
it's easy enough to fix.

Signed-off-by: Olivier Fourdan <ofourdan@redhat.com>
Reviewed-by: Peter Hutterer <peter.hutterer@who-t.net>
Reviewed-by: Michel Dänzer <mdaenzer@redhat.com>